requirements.txt
numpy
cupy-cuda11x
opencv-python
scikit-image
matplotlib
h5py
tqdm
scikit-learn